Adding more tests.
Notably, some tests are now run in separate processes so they can mutate global state (tracing) without impacting each other. A test was also added that runs the actual server, test it, and kill it.
