+++ crates/libmotiva/src/matching/mod.rs
@@ -1,5 +1,8 @@
 mod matchers;
 
+#[cfg(test)]
+mod tests;
+
 use bumpalo::Bump;
 use jiff::Timestamp;
 use serde::Deserialize;
