+++ crates/libmotiva/src/matching/tests/mod.rs
@@ -1,0 +1,1 @@
+mod comprehensive;
