+++ crates/motiva/Cargo.toml
@@ -67,5 +67,7 @@ axum-macros = "0.5.0"
 
 [dev-dependencies]
 axum-test = "18.0.2"
+nix = { version = "0.30.1", features = ["signal"] }
+rusty-fork = "0.3.1"
 serde-json-assert = "0.4.0"
 serial_test = "3.2.0"
